pyproject.toml
[tool.poetry]
name = "flex-test"
version = "0.1.0"
description = ""
authors = ["Doka <dokadair@gmail.com>"]
readme = "README.md"
packages = [{include = "flex_test"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
celery = "^5.5.1"
django-cors-headers = "^4.7.0"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
psycopg2 = "^2.9.10"
gunicorn = "^23.0.0"
djangorestframework-simplejwt = "^5.5.0"
python-decouple = "^3.8"
django-celery-beat = "^2.8.0"
psycopg = "^3.2.6"
requests = "^2.32.3"
redis = "^5.2.1"
drf-spectacular-sidecar = "^2025.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
